+++ server/src/main/java/com/host/Dto/UserDto.java
@@ -1,4 +1,4 @@
 package com.host.Dto;
 
-public record UserDto(int u_id,String family_name,String username,String password,String email,String phone_number,String dob,String profession, String gender) {
+public record UserDto(long u_id,String family_name,String username,String password,String email,String phone_number,String dob,String profession, String gender) {
 }
