+++ server/src/main/java/com/host/Repositories/DemandeMaterialRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DemandeMaterialRepo  extends JpaRepository<Demand_materiel,Integer> {
+public interface DemandeMaterialRepo  extends JpaRepository<Demand_materiel,Long> {
 }
