+++ server/src/main/java/com/host/Repositories/MaterialRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaterialRepo  extends JpaRepository<Material,Integer> {
+public interface MaterialRepo  extends JpaRepository<Material,Long> {
 }
